Refactoring: make ScoreTypeWithUnitTest (naturally) available to all ScoreTypes
